requirements.txt
langchain
langchain-google-genai
langchain-groq
langchain-community
langgraph 
python-dotenv 
langchain-chroma 
PyMuPDF
langgraph-checkpoint-sqlite
sentence-transformers
